allow rowClass to send row index
This will allow classes to be added positionally

Instead of based on just value which is currently done in the demos.

For instance, if I want a css class applied to the last row in the table. Or the first row in the table.
Currently you have to use a css pseudo class like .rdg-row:nth-child(14) to make the 14th row have a certain css behavior.